# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Business Radar API
 * OpenAPI spec version: 0.1.0
 */
export type ClearWorkbookResultContextState = typeof ClearWorkbookResultContextState[keyof typeof ClearWorkbookResultContextState];


export const ClearWorkbookResultContextState = {
  NO_CONTEXT: 'NO_CONTEXT',
} as const;

export interface ClearWorkbookResult {
  contextState: ClearWorkbookResultContextState;
  message: string;
}

export interface HealthStatus {
  status: string;
}

export type Severity = typeof Severity[keyof typeof Severity];


export const Severity = {
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  LOW: 'LOW',
} as const;

export interface Anomaly {
  id: string;
  type: string;
  title: string;
  severity: Severity;
  summary: string;
  impactAmount: number;
  affectedRecords: number;
  affectedCustomers?: number;
  systems: string[];
  firstDetected: string;
  evidenceCount: number;
  recommendedAction: string;
}

export type EvidenceRecordMetadata = {[key: string]: string};

export interface EvidenceRecord {
  id: string;
  system: string;
  recordType: string;
  recordId: string;
  label: string;
  detail: string;
  status: string;
  timestamp: string;
  /** @nullable */
  amount?: number | null;
  metadata?: EvidenceRecordMetadata;
}

export interface Investigation {
  anomaly: Anomaly;
  executiveExplanation: string;
  knownFacts: string[];
  uncertainties: string[];
  possibleCause: string;
  evidence: EvidenceRecord[];
  affectedOrders?: string[];
  affectedCustomers?: string[];
}

export type RadarSummaryContextState = typeof RadarSummaryContextState[keyof typeof RadarSummaryContextState];


export const RadarSummaryContextState = {
  NO_CONTEXT: 'NO_CONTEXT',
  IMPORTED_CONTEXT: 'IMPORTED_CONTEXT',
} as const;

export interface RadarSummary {
  contextState: RadarSummaryContextState;
  businessSignals: number;
  highRisk: number;
  potentialExposure: number;
  affectedCustomers: number;
  lastScan: string;
  topAnomalies: Anomaly[];
}

export interface ScanResult {
  summary: RadarSummary;
  scannedAt: string;
  message: string;
}

export interface CopilotQuestion {
  /** @minLength 1 */
  question: string;
}

export interface CopilotAnswer {
  question: string;
  answer: string;
  basis: string[];
}

export interface Metric {
  label: string;
  value: string;
}

export interface DataSource {
  id: string;
  name: string;
  status: string;
  description: string;
  recordCount: number;
  source: string;
  active: boolean;
  /** @nullable */
  lastImported?: string | null;
  metrics: Metric[];
}

export interface BusinessPulse {
  revenue: number;
  availableCash: number;
  expectedSettlements: number;
  inventoryValue: number;
  outstandingReceivables: number;
}

export type BusinessOverviewContextState = typeof BusinessOverviewContextState[keyof typeof BusinessOverviewContextState];


export const BusinessOverviewContextState = {
  NO_CONTEXT: 'NO_CONTEXT',
  IMPORTED_CONTEXT: 'IMPORTED_CONTEXT',
} as const;

export interface BusinessOverview {
  contextState: BusinessOverviewContextState;
  merchant: string;
  subtitle: string;
  pulse: BusinessPulse | null;
  attentionCount: number;
  signals: Anomaly[];
  razorpayHealth: Metric[];
  businessHealth: Metric[];
}

export type WorkbookSheetPreviewItem = { [key: string]: unknown };

export interface WorkbookSheet {
  name: string;
  normalizedType: string;
  rowCount: number;
  supported: boolean;
  preview: WorkbookSheetPreviewItem[];
}

export interface ImportedDataset {
  recordCount: number;
  aliases: string[];
  available: boolean;
}

export type WorkbookImportDatasets = {[key: string]: ImportedDataset};

export interface DataHealth {
  recognizedDatasets: number;
  totalRecords: number;
  relationshipsResolved: number;
  missingRequiredFields: number;
  missingRequiredIds: number;
  duplicateIds: number;
  malformedDates: number;
  missingAmounts: number;
  unsupportedSheets: number;
  warnings: string[];
}

export interface WorkbookImport {
  fileName: string;
  importedAt: string;
  status: string;
  totalRecords: number;
  detectedSources: string[];
  sheets: WorkbookSheet[];
  datasets: WorkbookImportDatasets;
  dataHealth: DataHealth;
}

export interface WorkbookUpload {
  file: Blob;
}

export type DataHubContextState = typeof DataHubContextState[keyof typeof DataHubContextState];


export const DataHubContextState = {
  NO_CONTEXT: 'NO_CONTEXT',
  IMPORTED_CONTEXT: 'IMPORTED_CONTEXT',
} as const;

export interface ContextSource {
  id: string;
  name: string;
  recordCount: number;
  source: string;
  active: boolean;
}

export interface DataHub {
  contextState: DataHubContextState;
  razorpay: DataSource[];
  business: DataSource[];
  businessData: WorkbookImport | null;
  supportedSheets: string[];
  contextSources: ContextSource[];
  totalRecords: number;
  recognizedDatasets: number;
  contextStatus: string;
  /** @nullable */
  lastUpdated: string | null;
  dataHealth: DataHealth | null;
  suggestedPrompts: string[];
}

export interface CopilotEvidence {
  source: string;
  dataset: string;
  records: number;
  recordIds: string[];
}

export type BusinessCopilotAnswerConfidence = typeof BusinessCopilotAnswerConfidence[keyof typeof BusinessCopilotAnswerConfidence];


export const BusinessCopilotAnswerConfidence = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface BusinessCopilotAnswer {
  verdict: string;
  summary: string;
  calculation: Metric[];
  evidence: CopilotEvidence[];
  action: string;
  confidence: BusinessCopilotAnswerConfidence;
  usedAi: boolean;
}