# lib/api-zod/src/generated/types/anomaly.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Business Radar API
 * OpenAPI spec version: 0.1.0
 */
import type { Severity } from './severity';

export interface Anomaly {
  id: string;
  type: string;
  title: string;
  severity: Severity;
  summary: string;
  impactAmount: number;
  affectedRecords: number;
  affectedCustomers?: number;
  systems: string[];
  firstDetected: string;
  evidenceCount: number;
  recommendedAction: string;
}