# lib/api-zod/src/generated/types/businessOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Business Radar API
 * OpenAPI spec version: 0.1.0
 */
import type { Anomaly } from './anomaly';
import type { BusinessOverviewContextState } from './businessOverviewContextState';
import type { BusinessPulse } from './businessPulse';
import type { Metric } from './metric';

export interface BusinessOverview {
  contextState: BusinessOverviewContextState;
  merchant: string;
  subtitle: string;
  pulse: BusinessPulse | null;
  attentionCount: number;
  signals: Anomaly[];
  razorpayHealth: Metric[];
  businessHealth: Metric[];
}