# lib/api-zod/src/generated/types/businessPulse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Business Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessPulse {
  revenue: number;
  availableCash: number;
  expectedSettlements: number;
  inventoryValue: number;
  outstandingReceivables: number;
}