# lib/api-zod/src/generated/types/contextSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Business Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface ContextSource {
  id: string;
  name: string;
  recordCount: number;
  source: string;
  active: boolean;
}