# lib/api-zod/src/generated/types/copilotEvidence.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Business Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface CopilotEvidence {
  source: string;
  dataset: string;
  records: number;
  recordIds: string[];
}