# lib/api-zod/src/generated/types/copilotQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Business Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface CopilotQuestion {
  /** @minLength 1 */
  question: string;
}